Address review comments-1
